'Alert'

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { View, FlatList, Text } from 'react-native';
+import { View, FlatList, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { baseUrl } from '../shared/baseUrl';
@@ -18,7 +18,24 @@ const Favorites = (props) => {
             {
                 text: 'Delete',
                 type: 'delete',
-                onPress: () => deleteFavorite(item.id)
+                onPress: () => {
+                    Alert.alert(
+                        'Delete Favorite?',
+                        `Are you sure you wish to delete the favorite dish ${item.name}?`,
+                        [
+                            { 
+                                text: 'Cancel', 
+                                onPress: () => console.log(`${item.name} was not Deleted`),
+                                style: 'cancel'
+                            },
+                            {
+                                text: 'OK',
+                                onPress: () => deleteFavorite(item.id)
+                            }
+                        ],
+                        { cancelable: false }
+                    )
+                }
             }
         ];
